Modularize
Convert the "Fizz", "Buzz" and "Integer" code into modules that are
loaded dynamically using a ServiceLoader.

It is possible to change the set of modules that are used by simply
adding the required classes into a new module folder and adding the
module to the META-INF/services/ service descriptor.
