--- src/main/java/com/seriouscompany/business/java/fizzbuzz/packagenamingpackage/impl/factories/BuzzStrategyFactory.java
@@ -1,15 +1,0 @@
-package com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.factories;
-
-import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.strategies.BuzzStrategy;
-import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.factories.IsEvenlyDivisibleStrategyFactory;
-import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.strategies.IsEvenlyDivisibleStrategy;
-
-public class BuzzStrategyFactory implements IsEvenlyDivisibleStrategyFactory {
-
-	@Override
-	public IsEvenlyDivisibleStrategy createIsEvenlyDivisibleStrategy() {
-		final BuzzStrategy myBuzzStrategy = new BuzzStrategy();
-		return myBuzzStrategy;
-	}
-
-}
