--- src/main/java/com/seriouscompany/business/java/fizzbuzz/packagenamingpackage/impl/factories/FizzStringPrinterFactory.java
@@ -1,15 +1,0 @@
-package com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.factories;
-
-import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.printers.FizzStringPrinter;
-import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.factories.StringPrinterFactory;
-import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.printers.StringPrinter;
-
-public class FizzStringPrinterFactory implements StringPrinterFactory {
-
-	@Override
-	public StringPrinter createStringPrinter() {
-		final StringPrinter myFizzStringPrinter = new FizzStringPrinter();
-		return myFizzStringPrinter;
-	}
-
-}
