--- src/main/java/com/seriouscompany/business/java/fizzbuzz/packagenamingpackage/impl/factories/FizzStringReturnerFactory.java
@@ -1,15 +1,0 @@
-package com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.factories;
-
-import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.stringreturners.FizzStringReturner;
-import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.factories.StringStringReturnerFactory;
-import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.stringreturners.StringStringReturner;
-
-public class FizzStringReturnerFactory implements StringStringReturnerFactory {
-
-	@Override
-	public StringStringReturner createStringStringReturner() {
-		final FizzStringReturner myFizzStringReturner = new FizzStringReturner();
-		return myFizzStringReturner;
-	}
-
-}
