--- src/main/java/com/seriouscompany/business/java/fizzbuzz/packagenamingpackage/impl/factories/IntegerIntegerStringReturnerFactory.java
@@ -1,16 +1,0 @@
-package com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.factories;
-
-import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.stringreturners.IntegerIntegerStringReturner;
-import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.factories.IntegerStringReturnerFactory;
-import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.stringreturners.IntegerStringReturner;
-
-public class IntegerIntegerStringReturnerFactory implements
-		IntegerStringReturnerFactory {
-
-	@Override
-	public IntegerStringReturner createIntegerStringReturner() {
-		final IntegerIntegerStringReturner myIntegerIntegerStringReturner = new IntegerIntegerStringReturner();
-		return myIntegerIntegerStringReturner;
-	}
-
-}
