+++ src/main/java/com/seriouscompany/business/java/fizzbuzz/packagenamingpackage/impl/modules/buzz/constants/BuzzStrategyConstants.java
@@ -1,0 +1,6 @@
+package com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.modules.buzz.constants;
+
+public class BuzzStrategyConstants
+{
+	public static final int BUZZ_INTEGER_CONSTANT_VALUE = 5;
+}
