+++ src/main/java/com/seriouscompany/business/java/fizzbuzz/packagenamingpackage/impl/modules/buzz/factories/BuzzStrategyFactory.java
@@ -1,0 +1,15 @@
+package com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.modules.buzz.factories;
+
+import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.modules.buzz.strategies.BuzzStrategy;
+import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.factories.IsEvenlyDivisibleStrategyFactory;
+import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.strategies.IsEvenlyDivisibleStrategy;
+
+public class BuzzStrategyFactory implements IsEvenlyDivisibleStrategyFactory {
+
+	@Override
+	public IsEvenlyDivisibleStrategy createIsEvenlyDivisibleStrategy() {
+		final BuzzStrategy myBuzzStrategy = new BuzzStrategy();
+		return myBuzzStrategy;
+	}
+
+}
