+++ src/main/java/com/seriouscompany/business/java/fizzbuzz/packagenamingpackage/impl/modules/buzz/printers/BuzzStringPrinter.java
@@ -1,0 +1,35 @@
+package com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.modules.buzz.printers;
+
+import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.modules.buzz.factories.BuzzStringReturnerFactory;
+import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.factories.SystemOutFizzBuzzOutputStrategyFactory;
+import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.factories.FizzBuzzOutputStrategyFactory;
+import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.factories.StringStringReturnerFactory;
+import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.printers.StringPrinter;
+import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.strategies.FizzBuzzOutputStrategy;
+import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.stringreturners.StringStringReturner;
+import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.strategies.adapters.FizzBuzzOutputStrategyToFizzBuzzExceptionSafeOutputStrategyAdapter;
+
+public class BuzzStringPrinter implements StringPrinter {
+
+	private final FizzBuzzOutputStrategy outputStrategy;
+
+	public BuzzStringPrinter() {
+		FizzBuzzOutputStrategyFactory factory = new SystemOutFizzBuzzOutputStrategyFactory();
+		this.outputStrategy = factory.createOutputStrategy();
+	}
+
+	public void print() {
+		final StringStringReturnerFactory myBuzzStringReturnerFactory = new BuzzStringReturnerFactory();
+		final StringStringReturner myBuzzStringReturner = myBuzzStringReturnerFactory
+				.createStringStringReturner();
+		final FizzBuzzOutputStrategyToFizzBuzzExceptionSafeOutputStrategyAdapter myOutputAdapter =
+				new FizzBuzzOutputStrategyToFizzBuzzExceptionSafeOutputStrategyAdapter(outputStrategy);
+
+		myOutputAdapter.output(myBuzzStringReturner.getReturnString());
+	}
+
+	@Override
+	public void printValue(Object value) {
+		print();
+	}
+}
