+++ src/main/java/com/seriouscompany/business/java/fizzbuzz/packagenamingpackage/impl/modules/fizz/constants/FizzStrategyConstants.java
@@ -1,0 +1,6 @@
+package com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.modules.fizz.constants;
+
+public class FizzStrategyConstants
+{
+	public static final int FIZZ_INTEGER_CONSTANT_VALUE = 3;
+}
