+++ src/main/java/com/seriouscompany/business/java/fizzbuzz/packagenamingpackage/impl/modules/fizz/printers/FizzPrinter.java
@@ -1,0 +1,26 @@
+package com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.modules.fizz.printers;
+
+import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.factories.SystemOutFizzBuzzOutputStrategyFactory;
+import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.modules.fizz.returners.FizzStringReturner;
+import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.factories.FizzBuzzOutputStrategyFactory;
+import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.strategies.FizzBuzzOutputStrategy;
+import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.strategies.adapters.FizzBuzzOutputStrategyToFizzBuzzExceptionSafeOutputStrategyAdapter;
+
+public class FizzPrinter {
+
+	private final FizzBuzzOutputStrategy outputStrategy;
+
+	public FizzPrinter() {
+		FizzBuzzOutputStrategyFactory factory = new SystemOutFizzBuzzOutputStrategyFactory();
+		this.outputStrategy = factory.createOutputStrategy();
+	}
+
+	public void printFizz() {
+		final FizzStringReturner myFizzStringReturner = new FizzStringReturner();
+		final FizzBuzzOutputStrategyToFizzBuzzExceptionSafeOutputStrategyAdapter myOutputAdapter =
+				new FizzBuzzOutputStrategyToFizzBuzzExceptionSafeOutputStrategyAdapter(outputStrategy);
+
+		myOutputAdapter.output(myFizzStringReturner.getReturnString());
+	}
+
+}
