+++ src/main/java/com/seriouscompany/business/java/fizzbuzz/packagenamingpackage/impl/modules/nofizznobuzz/constants/NoFizzNoBuzzStrategyConstants.java
@@ -1,0 +1,7 @@
+package com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.modules.nofizznobuzz.constants;
+
+public class NoFizzNoBuzzStrategyConstants
+{
+	public static final int NO_BUZZ_INTEGER_CONSTANT_VALUE = 5;
+	public static final int NO_FIZZ_INTEGER_CONSTANT_VALUE = 3;
+}
