+++ src/main/java/com/seriouscompany/business/java/fizzbuzz/packagenamingpackage/impl/modules/nofizznobuzz/factories/NoFizzNoBuzzStrategyFactory.java
@@ -1,0 +1,16 @@
+package com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.modules.nofizznobuzz.factories;
+
+import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.modules.nofizznobuzz.strategies.NoFizzNoBuzzStrategy;
+import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.strategies.IsEvenlyDivisibleStrategy;
+import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.factories.IsEvenlyDivisibleStrategyFactory;
+
+public class NoFizzNoBuzzStrategyFactory implements
+		IsEvenlyDivisibleStrategyFactory {
+
+	@Override
+	public IsEvenlyDivisibleStrategy createIsEvenlyDivisibleStrategy() {
+		final NoFizzNoBuzzStrategy myNoFizzNoBuzzStrategy = new NoFizzNoBuzzStrategy();
+		return myNoFizzNoBuzzStrategy;
+	}
+
+}
