--- src/main/java/com/seriouscompany/business/java/fizzbuzz/packagenamingpackage/impl/printers/IntPrinter.java
@@ -1,29 +1,0 @@
-package com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.printers;
-
-import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.factories.SystemOutFizzBuzzOutputStrategyFactory;
-import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.stringreturners.IntegerIntegerStringReturner;
-import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.factories.FizzBuzzOutputStrategyFactory;
-import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.strategies.FizzBuzzOutputStrategy;
-import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.stringreturners.IntegerStringReturner;
-import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.strategies.adapters.FizzBuzzOutputStrategyToFizzBuzzExceptionSafeOutputStrategyAdapter;
-
-public class IntPrinter {
-
-	private final FizzBuzzOutputStrategy outputStrategy;
-
-	public IntPrinter() {
-		FizzBuzzOutputStrategyFactory factory = new SystemOutFizzBuzzOutputStrategyFactory();
-		this.outputStrategy = factory.createOutputStrategy();
-	}
-
-	public void printInteger(int theInteger) {
-		final IntegerStringReturner myIntegerIntegerStringReturner = new IntegerIntegerStringReturner();
-		final String myIntegerString = myIntegerIntegerStringReturner
-				.getIntegerReturnString(theInteger);
-		final FizzBuzzOutputStrategyToFizzBuzzExceptionSafeOutputStrategyAdapter myOutputAdapter =
-				new FizzBuzzOutputStrategyToFizzBuzzExceptionSafeOutputStrategyAdapter(outputStrategy);
-
-		myOutputAdapter.output(myIntegerString);
-	}
-
-}
