+++ src/main/java/com/seriouscompany/business/java/fizzbuzz/packagenamingpackage/impl/strategies/SingleStepOutputGenerationStrategy.java
@@ -4,18 +4,15 @@ import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 
-import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.factories.BuzzStrategyFactory;
-import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.factories.BuzzStringPrinterFactory;
+import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.modules.ModuleRegistry;
 import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.factories.FizzBuzzOutputGenerationContextVisitorFactory;
-import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.factories.FizzStrategyFactory;
-import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.factories.FizzStringPrinterFactory;
-import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.factories.IntIntegerPrinterFactory;
 import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.factories.NewLineStringPrinterFactory;
-import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.factories.NoFizzNoBuzzStrategyFactory;
 import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.visitors.FizzBuzzOutputGenerationContext;
+import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.modules.Module;
 import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.factories.IntegerPrinterFactory;
 import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.factories.IsEvenlyDivisibleStrategyFactory;
 import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.factories.OutputGenerationContextVisitorFactory;
+import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.factories.DataPrinterFactory;
 import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.factories.StringPrinterFactory;
 import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.printers.StringPrinter;
 import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.strategies.SingleStepOutputGenerationParameter;
@@ -30,22 +27,18 @@ public class SingleStepOutputGenerationStrategy {
 	private StringPrinter myNewLinePrinter;
 
 	public SingleStepOutputGenerationStrategy() {
+        for (Module module : ModuleRegistry.getModules()) {
+            final IsEvenlyDivisibleStrategyFactory myStrategyFactory = module.getStrategyFactory();
+            final DataPrinterFactory myPrinterFactory = module.getPrinterFactory();
+            contexts.add(new FizzBuzzOutputGenerationContext(myStrategyFactory.createIsEvenlyDivisibleStrategy(), 
+                                                             myPrinterFactory.createPrinter()));
+        }
+
 		OutputGenerationContextVisitorFactory contextVisitorFactory = new FizzBuzzOutputGenerationContextVisitorFactory();
-		contextVisitor = contextVisitorFactory.createVisitor(); 
-		final IsEvenlyDivisibleStrategyFactory myFizzStrategyFactory = new FizzStrategyFactory();
-		final StringPrinterFactory myFizzStringPrinterFactory = new FizzStringPrinterFactory();
-		contexts.add(new FizzBuzzOutputGenerationContext(myFizzStrategyFactory.createIsEvenlyDivisibleStrategy(), myFizzStringPrinterFactory.createStringPrinter()));
-
-		final IsEvenlyDivisibleStrategyFactory myBuzzStrategyFactory = new BuzzStrategyFactory();
-		final StringPrinterFactory myBuzzStringPrinterFactory = new BuzzStringPrinterFactory();
-		contexts.add(new FizzBuzzOutputGenerationContext(myBuzzStrategyFactory.createIsEvenlyDivisibleStrategy(), myBuzzStringPrinterFactory.createStringPrinter()));
-
-		final IsEvenlyDivisibleStrategyFactory myNoFizzNoBuzzStrategyFactory = new NoFizzNoBuzzStrategyFactory();
-		final IntegerPrinterFactory myIntIntegerPrinterFactory = new IntIntegerPrinterFactory();
-		contexts.add(new FizzBuzzOutputGenerationContext(myNoFizzNoBuzzStrategyFactory.createIsEvenlyDivisibleStrategy(), myIntIntegerPrinterFactory.createPrinter()));
+		contextVisitor = contextVisitorFactory.createVisitor();
 
 		final StringPrinterFactory myNewLineStringPrinterFactory = new NewLineStringPrinterFactory();
-		myNewLinePrinter = myNewLineStringPrinterFactory.createStringPrinter();
+		myNewLinePrinter = myNewLineStringPrinterFactory.createPrinter();
     }
 
 	public void performGenerationForCurrentStep(SingleStepOutputGenerationParameter generationParameter) {
