--- src/main/java/com/seriouscompany/business/java/fizzbuzz/packagenamingpackage/impl/strategies/constants/BuzzStrategyConstants.java
@@ -1,6 +1,0 @@
-package com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.strategies.constants;
-
-public class BuzzStrategyConstants
-{
-	public static final int BUZZ_INTEGER_CONSTANT_VALUE = 5;
-}
