--- src/main/java/com/seriouscompany/business/java/fizzbuzz/packagenamingpackage/impl/strategies/constants/FizzStrategyConstants.java
@@ -1,6 +1,0 @@
-package com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.strategies.constants;
-
-public class FizzStrategyConstants
-{
-	public static final int FIZZ_INTEGER_CONSTANT_VALUE = 3;
-}
