+++ src/main/java/com/seriouscompany/business/java/fizzbuzz/packagenamingpackage/impl/strategies/constants/NewLineStrategyConstants.java
@@ -1,0 +1,6 @@
+package com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.strategies.constants;
+
+public class NewLineStrategyConstants
+{
+	public static final int NEW_LINE_INTEGER_CONSTANT_VALUE = 1;
+}
