--- src/main/java/com/seriouscompany/business/java/fizzbuzz/packagenamingpackage/impl/strategies/constants/NoFizzNoBuzzStrategyConstants.java
@@ -1,7 +1,0 @@
-package com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.strategies.constants;
-
-public class NoFizzNoBuzzStrategyConstants
-{
-	public static final int NO_BUZZ_INTEGER_CONSTANT_VALUE = BuzzStrategyConstants.BUZZ_INTEGER_CONSTANT_VALUE;
-	public static final int NO_FIZZ_INTEGER_CONSTANT_VALUE = FizzStrategyConstants.FIZZ_INTEGER_CONSTANT_VALUE;
-}
