+++ src/main/java/com/seriouscompany/business/java/fizzbuzz/packagenamingpackage/interfaces/factories/DataPrinterFactory.java
@@ -1,0 +1,9 @@
+package com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.factories;
+
+import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.printers.DataPrinter;
+
+public interface DataPrinterFactory {
+
+	public DataPrinter createPrinter();
+
+}
