+++ src/main/java/com/seriouscompany/business/java/fizzbuzz/packagenamingpackage/interfaces/factories/IntegerPrinterFactory.java
@@ -2,8 +2,9 @@ package com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interface
 
 import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.printers.IntegerPrinter;
 
-public interface IntegerPrinterFactory {
-	
+public interface IntegerPrinterFactory extends DataPrinterFactory {
+
+    @Override
 	public IntegerPrinter createPrinter();
-	
+
 }
