+++ src/main/java/com/seriouscompany/business/java/fizzbuzz/packagenamingpackage/interfaces/factories/StringPrinterFactory.java
@@ -2,8 +2,9 @@ package com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interface
 
 import com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.printers.StringPrinter;
 
-public interface StringPrinterFactory {
-	
-	public StringPrinter createStringPrinter();
-	
+public interface StringPrinterFactory extends DataPrinterFactory {
+
+    @Override
+	public StringPrinter createPrinter();
+
 }
