+++ src/main/resources/META-INF/services/com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.interfaces.modules.Module
@@ -1,0 +1,3 @@
+com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.modules.fizz.FizzModule
+com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.modules.buzz.BuzzModule
+com.seriouscompany.business.java.fizzbuzz.packagenamingpackage.impl.modules.nofizznobuzz.NoFizzNoBuzzModule
